refactor(firebase): move firebase pull/push logic inside firebase dir
